int64

+++ rss/item_enclosure.go
@@ -3,6 +3,6 @@ package rss
 type ItemEnclosure struct {
 	URL  string `xml:"url,attr"`
 	Type string `xml:"type,attr"`
-	Length int  `xml:"length,attr"`
+	Length int64 `xml:"length,attr"`
 }
